requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-generator==1.10
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.11.2
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.54.0
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
h11==0.14.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
mssql-django==1.2
oauthlib==3.2.2
oscrypto==1.3.0
outcome==1.2.0
packaging==23.1
pdfminer.six==20221105
pdfplumber==0.8.0
pem==21.2.0
Pillow==9.4.0
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.6.0
pyodbc==4.0.35
pyparsing==3.0.9
pypdf==3.9.1
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
supervisor==4.1.0
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2022.7
tzlocal==5.0.1
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
Wand==0.6.11
webdriver-manager==3.8.6
webencodings==0.5.1
wfastcgi==3.0.0
wsproto==1.2.0
channels
daphne
channels_redis